fix count-safety issues in intranode bcast
MPIR_Get_count_impl is not count-safe.  because it is used to get the
byte count inside of intranode bcast, it will overflow even for counts
smaller than INT_MAX if type_size>1.

MPIR_Get_elements_x_impl is count-safe and provides the same result for
the case in question, as far as I can tell.

the issue was reported in
https://trac.mpich.org/projects/mpich/ticket/2311
the bug reproducer code attached to that ticket was used to verify the
correctness of the proposed solution.

a resolution is proposed in
https://trac.mpich.org/projects/mpich/ticket/2320#comment:1